+++ skrub/_dataframe/_common.py
@@ -85,6 +85,7 @@ __all__ = [
     "sample",
     "head",
     "replace",
+    "with_columns",
 ]
 
 #
@@ -1007,3 +1008,9 @@ def _replace_pandas(col, old, new):
 @replace.specialize("polars", argument_type="Column")
 def _replace_polars(col, old, new):
     return col.replace(old, new)
+
+
+def with_columns(df, **new_cols):
+    cols = {col_name: col(df, col_name) for col_name in column_names(df)}
+    cols.update({n: make_column_like(df, c, n) for n, c in new_cols.items()})
+    return make_dataframe_like(df, cols)