+++ skrub/tests/test_fuzzy_join.py
@@ -1,14 +1,13 @@
 import warnings
 
 import numpy as np
-import pandas as pd
 import pytest
 from numpy.testing import assert_array_equal
 from sklearn.feature_extraction.text import HashingVectorizer
 
-from skrub import fuzzy_join
+from skrub import ToDatetime, _join_utils, fuzzy_join
+from skrub import _selectors as s
 from skrub._dataframe import _common as ns
-from skrub._dataframe._testing_utils import assert_frame_equal
 
 
 @pytest.mark.parametrize(
@@ -19,8 +18,6 @@ def test_fuzzy_join(df_module, analyzer):
     """
     Testing if ``fuzzy_join`` results are as expected.
     """
-    if df_module.name == "polars":
-        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'.")
     df1 = df_module.make_dataframe({"a1": ["ana", "lala", "nana et sana", np.nan]})
     df2 = df_module.make_dataframe(
         {"a2": ["anna", "lala et nana", "lana", "sana", np.nan]}
@@ -50,16 +47,13 @@ def test_fuzzy_join(df_module, analyzer):
     # Joining is always done on the left table and thus takes it shape:
     assert ns.shape(df_joined2) == (len(df2), n_cols)
 
-    # TODO: dispatch ``with_columns``
-    df1["a2"] = 1
+    df1 = ns.with_columns(df1, **{"a2": [1] * ns.shape(df1)[0]})
 
     df_on = fuzzy_join(df_joined, df1, on="a1", suffix="2")
     assert "a12" in ns.column_names(df_on)
 
 
 def test_max_dist(df_module):
-    if df_module.name == "polars":
-        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'.")
     left = df_module.make_dataframe({"A": ["aa", "bb"]})
     right = df_module.make_dataframe({"A": ["aa", "ba"], "B": [1, 2]})
     join = fuzzy_join(left, right, on="A", suffix="r")
@@ -69,8 +63,6 @@ def test_max_dist(df_module):
 
 
 def test_perfect_matches(df_module):
-    if df_module.name == "polars":
-        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'.")
     # non-regression test for https://github.com/skrub-data/skrub/issues/764
     # fuzzy_join when all rows had a perfect match used to trigger a division by 0
     df = df_module.make_dataframe({"A": [0, 1]})
@@ -87,19 +79,15 @@ def test_fuzzy_join_dtypes(df_module):
     """
     Test that the dtypes of dataframes are maintained after join.
     """
-    if df_module.name == "polars":
-        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'.")
     a = df_module.make_dataframe({"col1": ["aaa", "bbb"], "col2": [1, 2]})
     b = df_module.make_dataframe({"col1": ["aaa_", "bbb_"], "col3": [1, 2]})
     c = fuzzy_join(a, b, on="col1", suffix="r")
-    assert ns.dtype(ns.col(a, "col2")).kind == "i"
+    assert ns.is_integer(ns.col(a, "col2"))
     assert ns.dtype(ns.col(c, "col2")) == ns.dtype(ns.col(a, "col2"))
     assert ns.dtype(ns.col(c, "col3r")) == ns.dtype(ns.col(b, "col3"))
 
 
 def test_missing_keys(df_module):
-    if df_module.name == "polars":
-        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'.")
     a = df_module.make_dataframe({"col1": ["aaa", "bbb"], "col2": [1, 2]})
     b = df_module.make_dataframe({"col1": ["aaa_", "bbb_"], "col3": [1, 2]})
     with pytest.raises(
@@ -116,8 +104,6 @@ def test_missing_keys(df_module):
 
 
 def test_drop_unmatched(df_module):
-    if df_module.name == "polars":
-        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'.")
     a = df_module.make_dataframe({"col1": ["aaaa", "bbb", "ddd dd"], "col2": [1, 2, 3]})
     b = df_module.make_dataframe(
         {"col1": ["aaa_", "bbb_", "cc ccc"], "col3": [1, 2, 3]}
@@ -137,20 +123,20 @@ def test_drop_unmatched(df_module):
     assert sum(ns.is_null(ns.col(c2, "col3r"))) > 0
 
 
-def test_fuzzy_join_pandas_comparison():
+def test_fuzzy_join_exact_matches(df_module):
     """
-    Tests if fuzzy_join's output is as similar as
-    possible with `pandas.merge`.
+    Tests if fuzzy_join's output is the same as a normal left-join when there
+    are exact matches for all rows.
     """
-    left = pd.DataFrame(
+    left = df_module.make_dataframe(
         {
-            "key": ["K0", "K1", "K2", "K3"],
+            "key": ["K2", "K2", "K3", "K1"],
             "A": ["A0", "A1", "A2", "A3"],
             "B": ["B0", "B1", "B2", "B3"],
         }
     )
 
-    right = pd.DataFrame(
+    right = df_module.make_dataframe(
         {
             "key_": ["K0", "K1", "K2", "K3"],
             "C": ["C0", "C1", "C2", "C3"],
@@ -158,20 +144,28 @@ def test_fuzzy_join_pandas_comparison():
         }
     )
 
-    result = pd.merge(left, right, left_on="key", right_on="key_")
+    result = _join_utils.left_join(left, right, left_on="key", right_on="key_")
     result_fj = fuzzy_join(
         left, right, left_on="key", right_on="key_", add_match_info=False
     )
+    df_module.assert_column_equal(
+        ns.col(result_fj, "key_"), ns.rename(ns.col(result_fj, "key"), "key_")
+    )
+    # `_left_join` does a (non-fuzzy, regular) equijoin so it only keeps one of the
+    # join columns (keeping both would be redundant as they are identical due to exact
+    # matching) -- same as the default behavior of polars (coalesce=True) and pandas.
+    # `fuzzy_join` keeps both columns because they are not identical, only up to
+    # fuzziness, so keeping both is informative. So here we drop `key_` to compare the
+    # 2 resulting dataframes.
+    result_fj = s.select(result_fj, ~s.cols("key_"))
 
-    assert_frame_equal(result, result_fj)
+    df_module.assert_frame_equal(result, result_fj)
 
 
 def test_correct_encoder(df_module):
     """
     Test that the encoder error checking is working as intended.
     """
-    if df_module.name == "polars":
-        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'.")
 
     class TestVectorizer(HashingVectorizer):
         """
@@ -213,8 +207,6 @@ def test_numerical_column(df_module):
     """
     Testing that ``fuzzy_join`` works with numerical columns.
     """
-    if df_module.name == "polars":
-        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'.")
     left = df_module.make_dataframe({"str1": ["aa", "a", "bb"], "int": [10, 2, 5]})
     right = df_module.make_dataframe(
         {
@@ -249,20 +241,20 @@ def test_datetime_column(df_module):
     """
     Testing that ``fuzzy_join`` works with datetime columns.
     """
-    if df_module.name == "polars":
-        pytest.xfail(reason="Module 'polars' has no attribute 'to_datetime'.")
+
+    def to_dt(lst):
+        return ToDatetime().fit_transform(df_module.make_column("", lst))
+
     left = df_module.make_dataframe(
         {
             "str1": ["aa", "a", "bb"],
-            "date": df_module.module.to_datetime(
-                ["10/10/2022", "12/11/2021", "09/25/2011"]
-            ),
+            "date": to_dt(["10/10/2022", "12/11/2021", "09/25/2011"]),
         }
     )
     right = df_module.make_dataframe(
         {
             "str2": ["aa", "bb", "a", "cc", "dd"],
-            "date": df_module.module.to_datetime(
+            "date": to_dt(
                 ["09/10/2022", "12/24/2021", "09/25/2010", "11/05/2025", "02/21/2000"]
             ),
         }
@@ -273,16 +265,12 @@ def test_datetime_column(df_module):
     fj_time_expected = df_module.make_dataframe(
         {
             "str1": ["aa", "a", "bb"],
-            "date": df_module.module.to_datetime(
-                ["10/10/2022", "12/11/2021", "09/25/2011"]
-            ),
+            "date": to_dt(["10/10/2022", "12/11/2021", "09/25/2011"]),
             "str2r": ["aa", "bb", "a"],
-            "dater": df_module.module.to_datetime(
-                ["09/10/2022", "12/24/2021", "09/25/2010"]
-            ),
+            "dater": to_dt(["09/10/2022", "12/24/2021", "09/25/2010"]),
         }
     )
-    assert_frame_equal(fj_time, fj_time_expected)
+    df_module.assert_frame_equal(fj_time, fj_time_expected)
 
     n_cols = ns.shape(left)[1] + ns.shape(right)[1]
     n_samples = len(left)
@@ -308,17 +296,17 @@ def test_mixed_joins(df_module):
     """
     Test fuzzy joining on mixed and multiple column types.
     """
-    if df_module.name == "polars":
-        pytest.xfail(reason="Module 'polars' has no attribute 'to_datetime'")
+
+    def to_dt(lst):
+        return ToDatetime().fit_transform(df_module.make_column("", lst))
+
     left = df_module.make_dataframe(
         {
             "str1": ["Paris", "Paris", "Paris"],
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": df_module.module.to_datetime(
-                ["10/10/2022", "12/11/2021", "09/25/2011"]
-            ),
+            "date": to_dt(["10/10/2022", "12/11/2021", "09/25/2011"]),
         }
     )
     right = df_module.make_dataframe(
@@ -327,7 +315,7 @@ def test_mixed_joins(df_module):
             "str_2": ["TX", "FR", "GR Mytho", "cc", "dd"],
             "int1": [55, 6, 2, 15, 6],
             "int2": [554, 146, 32, 215, 612],
-            "date": df_module.module.to_datetime(
+            "date": to_dt(
                 ["09/10/2022", "12/24/2021", "09/25/2010", "11/05/2025", "02/21/2000"]
             ),
         }
@@ -344,19 +332,15 @@ def test_mixed_joins(df_module):
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": df_module.module.to_datetime(
-                ["10/10/2022", "12/11/2021", "09/25/2011"]
-            ),
+            "date": to_dt(["10/10/2022", "12/11/2021", "09/25/2011"]),
             "str_1r": ["Paris", "Paris", "dd"],
             "str_2r": ["FR", "FR", "dd"],
             "int1r": [6, 6, 6],
             "int2r": [146, 146, 612],
-            "dater": df_module.module.to_datetime(
-                ["12/24/2021", "12/24/2021", "02/21/2000"]
-            ),
+            "dater": to_dt(["12/24/2021", "12/24/2021", "02/21/2000"]),
         }
     )
-    assert_frame_equal(fj_num, expected_fj_num)
+    df_module.assert_frame_equal(fj_num, expected_fj_num)
     assert ns.shape(fj_num) == (3, 10)
 
     # On multiple string keys
@@ -375,19 +359,15 @@ def test_mixed_joins(df_module):
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": df_module.module.to_datetime(
-                ["2022-10-10", "2021-12-11", "2011-09-25"]
-            ),
+            "date": to_dt(["2022-10-10", "2021-12-11", "2011-09-25"]),
             "str_1r": ["Paris", "Paris", "Paris"],
             "str_2r": ["TX", "FR", "GR Mytho"],
             "int1r": [55, 6, 2],
             "int2r": [554, 146, 32],
-            "dater": df_module.module.to_datetime(
-                ["2022-09-10", "2021-12-24", "2010-09-25"]
-            ),
+            "dater": to_dt(["2022-09-10", "2021-12-24", "2010-09-25"]),
         }
     )
-    assert_frame_equal(fj_str, expected_fj_str)
+    df_module.assert_frame_equal(fj_str, expected_fj_str)
     assert ns.shape(fj_str) == (3, 10)
 
     # On mixed, numeric and string keys
@@ -405,19 +385,15 @@ def test_mixed_joins(df_module):
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": df_module.module.to_datetime(
-                ["2022-10-10", "2021-12-11", "2011-09-25"]
-            ),
+            "date": to_dt(["2022-10-10", "2021-12-11", "2011-09-25"]),
             "str_1r": ["Paris", "Paris", "Paris"],
             "str_2r": ["FR", "FR", "TX"],
             "int1r": [6, 6, 55],
             "int2r": [146, 146, 554],
-            "dater": df_module.module.to_datetime(
-                ["2021-12-24", "2021-12-24", "2022-09-10"]
-            ),
+            "dater": to_dt(["2021-12-24", "2021-12-24", "2022-09-10"]),
         }
     )
-    assert_frame_equal(fj_mixed, expected_fj_mixed)
+    df_module.assert_frame_equal(fj_mixed, expected_fj_mixed)
     assert ns.shape(fj_mixed) == (3, 10)
 
     # On mixed time and string keys
@@ -435,19 +411,15 @@ def test_mixed_joins(df_module):
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": df_module.module.to_datetime(
-                ["2022-10-10", "2021-12-11", "2011-09-25"]
-            ),
+            "date": to_dt(["2022-10-10", "2021-12-11", "2011-09-25"]),
             "str_1r": ["Paris", "Paris", "Paris"],
             "str_2r": ["TX", "FR", "GR Mytho"],
             "int1r": [55, 6, 2],
             "int2r": [554, 146, 32],
-            "dater": df_module.module.to_datetime(
-                ["2022-09-10", "2021-12-24", "2010-09-25"]
-            ),
+            "dater": to_dt(["2022-09-10", "2021-12-24", "2010-09-25"]),
         }
     )
-    assert_frame_equal(fj_mixed2, expected_fj_mixed2)
+    df_module.assert_frame_equal(fj_mixed2, expected_fj_mixed2)
     assert ns.shape(fj_mixed2) == (3, 10)
 
     # On mixed time and numbers keys
@@ -465,19 +437,15 @@ def test_mixed_joins(df_module):
             "str2": ["Texas", "France", "Greek God"],
             "int1": [10, 2, 5],
             "int2": [103, 250, 532],
-            "date": df_module.module.to_datetime(
-                ["2022-10-10", "2021-12-11", "2011-09-25"]
-            ),
+            "date": to_dt(["2022-10-10", "2021-12-11", "2011-09-25"]),
             "str_1r": ["Paris", "Paris", "Paris"],
             "str_2r": ["FR", "FR", "GR Mytho"],
             "int1r": [6, 6, 2],
             "int2r": [146, 146, 32],
-            "dater": df_module.module.to_datetime(
-                ["2021-12-24", "2021-12-24", "2010-09-25"]
-            ),
+            "dater": to_dt(["2021-12-24", "2021-12-24", "2010-09-25"]),
         }
     )
-    assert_frame_equal(fj_mixed3, expected_fj_mixed3)
+    df_module.assert_frame_equal(fj_mixed3, expected_fj_mixed3)
     assert ns.shape(fj_mixed3) == (3, 10)
 
 
@@ -485,8 +453,6 @@ def test_iterable_input(df_module):
     """
     Test if iterable inputs (list, set, dictionary or tuple) work.
     """
-    if df_module.name == "polars":
-        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'.")
     df1 = df_module.make_dataframe(
         {"a": ["ana", "lala", "nana"], "str2": ["Texas", "France", "Greek God"]}
     )
@@ -529,20 +495,15 @@ def test_iterable_input(df_module):
     ) == (3, 4)
 
 
-@pytest.mark.xfail
 def test_missing_values(df_module):
     """
     Test fuzzy joining on missing values.
     """
-    if df_module.name == "polars":
-        pytest.xfail(reason="Polars DataFrame object has no attribute 'reset_index'.")
     a = df_module.make_dataframe({"col1": ["aaaa", "bbb", "ddd dd"], "col2": [1, 2, 3]})
     b = df_module.make_dataframe({"col3": [np.nan, "bbb", "ddd dd"], "col4": [1, 2, 3]})
 
-    with pytest.warns(UserWarning, match=r"merging on missing values"):
-        c = fuzzy_join(a, b, left_on="col1", right_on="col3", add_match_info=False)
+    c = fuzzy_join(a, b, left_on="col1", right_on="col3", add_match_info=False)
     assert ns.shape(c)[0] == len(b)
 
-    with pytest.warns(UserWarning, match=r"merging on missing values"):
-        c = fuzzy_join(b, a, left_on="col3", right_on="col1", add_match_info=True)
+    c = fuzzy_join(b, a, left_on="col3", right_on="col1", add_match_info=True)
     assert ns.shape(c)[0] == len(b)