+++ skrub/tests/test_joiner.py
@@ -1,3 +1,5 @@
+import datetime
+
 import numpy as np
 import pandas as pd
 import pytest
@@ -5,7 +7,6 @@ from numpy.testing import assert_array_equal
 
 from skrub import Joiner
 from skrub._dataframe import _common as ns
-from skrub._dataframe._testing_utils import assert_frame_equal
 
 
 @pytest.fixture
@@ -49,22 +50,17 @@ def test_fit_transform(main_table, aux_table):
 
 
 def test_wrong_main_key(main_table, aux_table):
-    false_joiner = Joiner(aux_table=aux_table, main_key="Countryy", aux_key="country")
-
-    with pytest.raises(
-        ValueError,
-        match="do not exist in 'X'",
-    ):
-        false_joiner.fit(main_table)
+    wrong_joiner = Joiner(aux_table=aux_table, main_key="wrong_key", aux_key="country")
+    with pytest.raises(ValueError, match="do not exist in 'X'"):
+        wrong_joiner.fit(main_table)
 
 
 def test_wrong_aux_key(main_table, aux_table):
-    false_joiner2 = Joiner(aux_table=aux_table, main_key="Country", aux_key="bad")
-    with pytest.raises(
-        ValueError,
-        match="do not exist in 'aux_table'",
-    ):
-        false_joiner2.fit(main_table)
+    wrong_joiner_2 = Joiner(
+        aux_table=aux_table, main_key="Country", aux_key="wrong_key"
+    )
+    with pytest.raises(ValueError, match="do not exist in 'aux_table'"):
+        wrong_joiner_2.fit(main_table)
 
 
 def test_multiple_keys(df_module):
@@ -76,20 +72,20 @@ def test_multiple_keys(df_module):
     )
     joiner_list = Joiner(
         aux_table=df2,
-        aux_key=["CO", "CA"],
         main_key=["Co", "Ca"],
+        aux_key=["CO", "CA"],
         add_match_info=False,
     )
     result = joiner_list.fit_transform(df)
 
     expected = ns.concat_horizontal(df, df2)
-    assert_frame_equal(result, expected)
+    df_module.assert_frame_equal(result, expected)
 
     joiner_list = Joiner(
-        aux_table=df2, aux_key="CA", main_key="Ca", add_match_info=False
+        aux_table=df2, main_key="Ca", aux_key="CA", add_match_info=False
     )
     result = joiner_list.fit_transform(df)
-    assert_frame_equal(result, expected)
+    df_module.assert_frame_equal(result, expected)
 
 
 def test_pandas_aux_table_index():
@@ -108,22 +104,54 @@ def test_pandas_aux_table_index():
         suffix="_capitals",
     )
     join = joiner.fit_transform(main_table)
-    assert ns.to_list(ns.col(join, "Country_capitals")) == [
+    assert join["Country_capitals"].to_list() == [
         "France",
         "Italy",
         "Germany",
     ]
 
 
-def test_bad_ref_dist():
-    table = pd.DataFrame({"A": [1, 2]})
-    joiner = Joiner(table, key="A", ref_dist="bad")
-    with pytest.raises(ValueError, match="got 'bad'"):
-        joiner.fit(table)
+def test_wrong_ref_dist(main_table, aux_table):
+    joiner = Joiner(
+        aux_table, main_key="Country", aux_key="country", ref_dist="wrong_ref_dist"
+    )
+    with pytest.raises(
+        ValueError, match=r"'ref_dist' should be one of.* Got 'wrong_ref_dist'"
+    ):
+        joiner.fit(main_table)
 
 
 @pytest.mark.parametrize("max_dist", [np.inf, float("inf"), "inf", None])
-def test_max_dist(max_dist):
-    table = pd.DataFrame({"A": [1, 2]})
-    joiner = Joiner(table, key="A", max_dist=max_dist, suffix="_").fit(table)
+def test_max_dist(main_table, aux_table, max_dist):
+    joiner = Joiner(
+        aux_table, main_key="Country", aux_key="country", max_dist=max_dist
+    ).fit(main_table)
     assert joiner.max_dist_ == np.inf
+
+
+def test_missing_values(df_module):
+    df = df_module.make_dataframe({"A": [None, "hollywood", "beverly"]})
+    joiner = Joiner(df, key="A", suffix="_")
+    out = joiner.fit_transform(df)
+    assert ns.shape(out) == (3, 5)
+
+
+def test_fit_transform_numeric(df_module):
+    df = df_module.make_dataframe({"A": [4.5, 0.5, 1, -1.5]})
+    joiner = Joiner(df, key="A", suffix="_")
+    out = joiner.fit_transform(df)
+    assert ns.shape(out) == (4, 5)
+
+
+def test_fit_transform_datetimes(df_module):
+    values = [
+        datetime.datetime.fromisoformat(dt)
+        for dt in [
+            "2020-02-03T12:30:05",
+            "2021-03-15T00:37:15",
+            "2022-02-13T17:03:25",
+        ]
+    ]
+    df = df_module.make_dataframe({"A": values})
+    joiner = Joiner(df, key="A", suffix="_")
+    joiner.fit_transform(df)